// src/core/games/games.types.ts
export enum GameRevisionStatus {
  Draft = 'draft',
  Review = 'review',
  Published = 'published',
}
export enum GameAuthorType {
  Discord = 'discord',
  Text = 'text',
}
export enum GameAttachmentType {
  Image = 'image',
  ExternalVideo = 'external_video',
}
export enum GameReviewAction {
  Submitted = 'submitted',
  Published = 'published',
  ChangesRequested = 'changes_requested',
}
export enum GameListSort {
  ReleaseDateDesc = 'release_date_desc',
  ReleaseDateAsc = 'release_date_asc',
  LikesDesc = 'likes_desc',
  PublishedDesc = 'published_desc',
}